added options to the command line interface
Encoding utf-8 is ensured now.
Multiple input files can be passed.
The names of the crawlers and stores can be passed, matching with the input files.
It is not longer necessary to state the name of the table in the input file.
If the output file already exists, it will not be overwritten but used.